@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.12",
+    "@angular/animations": "18.2.13",
     "@angular/cdk": "15.1.3",
-    "@angular/common": "17.3.12",
-    "@angular/compiler": "17.3.12",
-    "@angular/core": "17.3.12",
-    "@angular/forms": "17.3.12",
-    "@angular/localize": "17.3.12",
-    "@angular/platform-browser": "17.3.12",
-    "@angular/platform-browser-dynamic": "17.3.12",
-    "@angular/platform-server": "17.3.12",
-    "@angular/router": "17.3.12",
+    "@angular/common": "18.2.13",
+    "@angular/compiler": "18.2.13",
+    "@angular/core": "18.2.13",
+    "@angular/forms": "18.2.13",
+    "@angular/localize": "18.2.13",
+    "@angular/platform-browser": "18.2.13",
+    "@angular/platform-browser-dynamic": "18.2.13",
+    "@angular/platform-server": "18.2.13",
+    "@angular/router": "18.2.13",
     "@fortawesome/fontawesome-free": "^6.1.1",
     "@ng-bootstrap/ng-bootstrap": "14.0.1",
     "@popperjs/core": "^2.11.5",
@@ -40,9 +40,9 @@
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.17",
-    "@angular/cli": "17.3.17",
-    "@angular/compiler-cli": "17.3.12",
+    "@angular/build": "^18.2.19",
+    "@angular/cli": "18.2.19",
+    "@angular/compiler-cli": "18.2.13",
     "typescript": "^5.4.5"
   }
 }
